+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and what APIs have changed, if applicable.
 ## [29.4.6] - 2020-07-29
 - Provide a default symbol table provider implementation that doesn't use symbol tables for requests/responses of its own, but is able to retrieve remote symbol tables to decode responses from other services (#357)
 - Provide public method in the AbstractRequestBuilder for adding field projections (#353)
+- Allow RestRestliServer and StreamRestliServer throw RestException & StreamException with no stacktrace
 
 ## [29.4.5] - 2020-07-21
 - Update ExtensionSchemaValidation task to check extension schema annotation (#254)